Make games/friends state available on client side immediately
Refactor the config and friends components and associated routes to ensure that friends data is available on the client side immediately after a new API key and Steam ID was posted.

Updates to the session store are only available to preload functions on components after a full refresh. Conditionally making a GET request to /api/friends or /api/key when the data is not in the session store means that we can get data in the client even without a refresh.